Fix for make links for broken ancient horrible /bin/sh

+++ Server/bin/make_links.sh
@@ -7,9 +7,20 @@ do
       for x in dbconvert mkhtml mkindx netrhost netrhost.debugmon
       do
 	 echo "     Creating Link: $x"|tr -d '\012'
+         if [ ! -f ../bin/$x ] 
+         then  
+            echo " unable to find binary!"
+            exit 1
+         fi
          rm $x 2>/dev/null
-         [[ ! -f ../bin/$x ]] && echo " unable to find binary!" && exit 1
-         [[ ! -h $x ]] && ln -s ../bin/$x $x 2>/dev/null || echo "..[ skipping ln ]"|tr -d '\012' 
+         if [ ! -h $x ] 
+         then
+            ln -s ../bin/$x $x 2>/dev/null 
+            if [ $? -ne 0 ]
+            then
+               echo "..[ skipping ln ]"|tr -d '\012' 
+            fi
+         fi
          echo " completed."
       done
       cd ..
